requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-genai>=1.0.0
python-dotenv==1.0.0
boto3==1.28.85
openai>=1.0.0
